The_HIVE/requirements.txt
openai>=1.10.0
sounddevice>=0.4.6
numpy>=1.24.3
websocket-client>=1.6.1
requests>=2.31.0
python-dotenv>=1.0.0
playsound==1.2.2
